feat: Added service-worker
Added all files required for a service worker to be produced.

The service worker is built using a post build script instead of being
an integral part of the angular build and serve process. This means that
in order to test the service worker, one has to run a production build
first and then run the "watch" script afterwards, which spins up the
express server and serves the complete application.

I would very much like to find a way to integrate this into the angular
build process instead, but I have yet to find a way to do this where I
can get access to the complete output from the build and not just the js
assets.