Fund widget (#16)
Trying out an integration to nordnet displaying a couple of funds. 

Also tyring out Apache Echarts for timeseries display. This brought on a bit of a challenge that I could not overcome. It seems that ECharts and Jest do not play nice together. Jest complained about imports from echarts having unexpected tokens. Searching the web it seems that many are having issues with this. Perhaps this is not the right tool for the job?