feat: Added option to use a database (#35)
This enables the backend code to utilize a small sqlite database for persistance. 

Backend can be run separately and frontend will auto-detect this and just instantiate an express instance for its SSR needs. But if backend is not active when frontend is started, the SSR instance will be a full blown NestJS application.